// src/app/interface/tipo.oraciones.interface.ts
// Generated by https://quicktype.io

export interface ResultTipoOraciones {
  ok:          boolean;
  msg:         string;
  tipooracion: Tipooracion[];
}

export interface ResultTipoOracion {
  ok:          boolean;
  msg:         string;
  tipooracion: Tipooracion;
}

export interface Tipooracion {
  id:             number;
  titulo:         string;
  descripcion:    string;
  subdescripcion: string;
  logo:           string;
  id_oracion:     number;
  Oracione:       Oracione;
}

export interface Oracione {
  id:       number;
  nombre:   string;
  multiple: number;
  logo:     string;
  estado:   number;
}

export interface TipoOracionesForm{
  titulo:string,
  descripcion:string,
  subdescripcion:string,
  logo?:string,
  id_oracion:string
}
export interface LogoTipoOracionesForm{
  logo?:string,
}