// src/app/interface/oracion.interface.ts
// Generated by https://quicktype.io
export interface OracionesForm{
  nombre?:string,
  multiple?:string,
  logo?:string
}

export interface ResultOraciones {
  ok:        boolean;
  msg:       string;
  oraciones: Oracione[];
}
export interface ResultOracion {
  ok:        boolean;
  msg:       string;
  oraciones: Oracione;
}

export interface Oracione {
  id:       number;
  nombre:   string;
  multiple: number;
  logo:     string;
  estado:   number;
}